fix(server): one `send`, paced, and the test that watches it loads again
`tests/library/serve.test.js` has not loaded since `send` moved into
`@uniflowed/server/node` and its import did not follow:

    ✗ serve.test.js failed to load: SyntaxError: The requested module
      '../../packages/vite/internal/serve.js' does not provide an export named 'send'

`uf test` printed `1564 passed, 0 failed` above `error: uf test could not run
1 file` — a red run that looks green in every line but the last, with nineteen
assertions about the two handlers quietly not being made.

And the file it could not load was guarding a fix that had been undone. The
`send` that moved lost the socket pacing and the hang-up cancel this branch
added three commits earlier, so `uf start` and every adapter buffered a slow
client's body in memory while `uf dev`, `uf preview` and the compiled binary
paced it. A deployment whose memory profile differs from the one that was
checked is exactly what the comment on the copy in `standalone.js` said must
not happen.

There is one copy now. `internal/http.js` re-exports it, and `standalone.js`
imports it — the reason that file gave for writing its own no longer holds,
because the loop lives in `@uniflowed/server`, which is the package it is in.
`HEAD` stays at its call site, being a decision about a request rather than
about writing a body.

Both guarantees are held by mutation: removing the back-pressure check fails
"stops reading while the socket is full, and goes on when it drains", and
removing the `close` listener fails that one and "cancels the body when the
client hangs up".

`uf test#library` 1583 passed, 0 failed — nineteen more than before, which are
the ones that were not running.

Closes #400
