+++ packages/vite/internal/http.js
@@ -55,80 +55,11 @@ export async function toRequest(incoming, config) {
 /**
  * Write a `Response` to a Node response.
  *
- * @param {import("node:http").ServerResponse} outgoing
- * @param {Response} result
+ * Re-exported rather than written again. This was a second copy of the loop
+ * in `@uniflowed/server`'s `node.js`, and the two drifted the moment the
+ * shared one moved: `uf start` and every adapter lost the socket pacing and
+ * the hang-up cancel while `uf dev` and `uf preview` kept them, which is the
+ * memory profile of a deployment differing from the one that was checked.
+ * See ubugeeei-prod/uf#400.
  */
-export async function send(outgoing, result) {
-  outgoing.statusCode = result.status;
-  if (result.statusText !== "") {
-    outgoing.statusMessage = result.statusText;
-  }
-  for (const [name, value] of result.headers) {
-    outgoing.setHeader(name, value);
-  }
-  if (result.body == null) {
-    outgoing.end();
-    return;
-  }
-  // Streamed rather than buffered, so a handler returning a large or
-  // open-ended body is not read into memory first.
-  //
-  // Which was only half true while this loop read as fast as the body would
-  // give: `write` answers `false` when the kernel buffer is full and the rest
-  // is being held in *this process's* memory, and a reader that ignores that
-  // turns a slow client into a heap the size of everything it has not
-  // acknowledged. Streaming in shape and buffering in fact — the same failure
-  // `ChunkQueue` exists to avoid a layer up, in the renderer.
-  const reader = result.body.getReader();
-  // And a client that hangs up is the other half. Nothing written after that
-  // goes anywhere, and the producer behind the body — a render, a proxied
-  // upstream, an event stream — keeps producing for a reader that is never
-  // coming back. `cancel()` is what tells it to stop; `releaseLock()` would
-  // only detach this end.
-  let open = true;
-  const onClose = () => {
-    open = false;
-  };
-  outgoing.on("close", onClose);
-  try {
-    while (open) {
-      const { done, value } = await reader.read();
-      if (done || !open) break;
-      if (outgoing.write(value) === false) {
-        await writable(outgoing);
-      }
-    }
-  } finally {
-    outgoing.off("close", onClose);
-  }
-  if (open) {
-    outgoing.end();
-    return;
-  }
-  // Best effort, and the only place in this file where a rejection is dropped:
-  // the connection is already gone, so there is nobody left to report to and
-  // no response left to fail.
-  await reader.cancel().catch(() => {});
-}
-
-/**
- * Resolve once `outgoing` can take more — or once it cannot ever again.
- *
- * `drain` alone would be a deadlock waiting to happen: a client that hangs up
- * while the buffer is full emits `close` and never `drain`, and a writer
- * waiting only for the latter waits for the life of the process, holding the
- * body's producer open with it.
- *
- * @param {import("node:http").ServerResponse} outgoing
- */
-function writable(outgoing) {
-  return new Promise((resolve) => {
-    const settle = () => {
-      outgoing.off("drain", settle);
-      outgoing.off("close", settle);
-      resolve();
-    };
-    outgoing.once("drain", settle);
-    outgoing.once("close", settle);
-  });
-}
+export { send } from "@uniflowed/server/node";
